Adjust Ray t0 to be optical path length
This should enable production of optical path differences.

+++ pysrc/vec3.cpp
@@ -51,10 +51,10 @@ namespace jtrace {
             .def(py::self += py::self)
             .def(py::self - py::self)
             .def(py::self -= py::self)
-            .def(py::self * float())
-            .def(py::self *= float())
-            .def(py::self / float())
-            .def(py::self /= float())
+            .def(py::self * double())
+            .def(py::self *= double())
+            .def(py::self / double())
+            .def(py::self /= double())
             .def(py::self == py::self)
             .def(py::self != py::self)
             .def(-py::self);
@@ -73,10 +73,10 @@ namespace jtrace {
             .def(py::init<>())
             .def(py::init<std::array<double,9>>())
             .def("__repr__", &Rot3::repr)
-            .def(py::self * float())
-            .def(py::self *= float())
-            .def(py::self / float())
-            .def(py::self /= float())
+            .def(py::self * double())
+            .def(py::self *= double())
+            .def(py::self / double())
+            .def(py::self /= double())
             .def("determinant", &Rot3::determinant);
 
         m.def("DotProduct", &DotProduct, R"pbdoc(
